ArmVirtPkg/QemuVirtMemInfoLib: Initialize all MMIO regions for CCA
A Realm guest needs to access peripherals emulated with the host with
the NS bit set (most significant bit of the Intermediate Physical
Address). Initialize the page tables with all device mappings. We mainly
need the serial device to be immediately accessible for debugging
purposes, but we may as well setup the whole device regions.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>