ArmVirtPkg: Kvmtool - Parse FDT to populate mem map
The memory map for the peripherals available for a Guest VM
launched using Kvmtool is currently populated by treating
all the memory before and after the System DRAM as device
memory.

With the introduction of ArmPlatfromDeviceInfoLib it is
now possible to get the MMIO base address and range for
the devices that are present in the FDT.

Therefore, use the ArmPlatfromDeviceInfoLib to get the
PLATFROM_DEVICE_INFO and populate the mem map accordingly.

Cc: Ard Biesheuvel <ardb+tianocore@kernel.org>
Cc: Leif Lindholm <quic_llindhol@quicinc.com>
Cc: Gerd Hoffmann <kraxel@redhat.com>
Signed-off-by: Sami Mujawar <sami.mujawar@arm.com>